option pour ne pas supprimer le fichier initial
Option -n ajouter pour empêcher la suppression du fichier initial.

+++ codec.sh
@@ -113,7 +113,8 @@ force=0
 v_copy=0
 thread=0
 deleteDuplicate=0
-while getopts "i:t:f:cd" option
+no_copy=0
+while getopts "i:t:f:cdn" option
 do
     case $option in
     i)
@@ -143,6 +144,9 @@ do
 	    c)
 		    v_copy=1
 		    ;;
+            n)
+	            no_copy=1
+                    ;;
 	    :)
 		    echo "L'option $OPTARG requiert un argument"
 		    ;;
@@ -394,14 +398,19 @@ do
                         # sinon on supprime le fichier de base devenu inutile
                         if [ "$same" == "1" ]
                         then
-                            hist=`grep -rne "$init" ~/.encode_file | cut -f1 -d ':'`
-                            if [ "$hist" != "" ]
+                            if [ "$n" == "0" ] 
                             then
-                                sed -i $hist'd' ~/.encode_file
+                                hist=`grep -rne "$init" ~/.encode_file | cut -f1 -d ':'`
+                                if [ "$hist" != "" ]
+                                then
+                                    sed -i $hist'd' ~/.encode_file
+                                fi
+                                echo "$init#$file_encode_txt#$hd" >> ~/.encode_file
+                                echo "mv $to $init"
+                                mv "$to" "$init"
+                            else
+                                echo "$to#$file_encode_txt#$hd" >> ~/.encode_file
                             fi
-                            echo "$init#$file_encode_txt#$hd" >> ~/.encode_file
-                            echo "mv $to $init"
-                            mv "$to" "$init"
                         else
                             hist=`grep -rne "$to" ~/.encode_file | cut -f1 -d ':'`
                             if [ "$hist" != "" ]
@@ -409,8 +418,11 @@ do
                                 sed -i $hist'd' ~/.encode_file
                             fi
                             echo "$to#$file_encode_txt#$hd" >> ~/.encode_file
-                            echo "rm $init"
-                            rm "$init"
+                            if [ "$n" == "0" ]
+                            then
+                                echo "rm $init"
+                                rm "$init"
+                            fi
                         fi
                         notify-send "convertion de $init terminée en $runtime secondes"
                     else
@@ -464,14 +476,19 @@ do
                                 # sinon on supprime le fichier de base devenu inutile
                                 if [ "$same" == "1" ]
                                 then
-                                    hist=`grep -rne "$init" ~/.encode_file | cut -f1 -d ':'`
-                                    if [ "$hist" != "" ]
-                                    then
-                                        sed -i $hist'd' ~/.encode_file
+                                  
+                                    if [ "$n" == "0" ] 
+                                        hist=`grep -rne "$init" ~/.encode_file | cut -f1 -d ':'`
+                                        if [ "$hist" != "" ]
+                                        then
+                                            sed -i $hist'd' ~/.encode_file
+                                        fi
+                                        echo "$init#$file_encode_txt#$hd" >> ~/.encode_file
+                                        echo "mv $to $init"
+                                        mv "$to" "$init"
+                                    else
+                                        echo "$to#$file_encode_txt#$hd" >> ~/.encode_file
                                     fi
-                                    echo "$init#$file_encode_txt#$hd" >> ~/.encode_file
-                                    echo "mv $to $init"
-                                    mv "$to" "$init"
                                 else
                                     hist=`grep -rne "$to" ~/.encode_file | cut -f1 -d ':'`
                                     if [ "$hist" != "" ]
@@ -479,8 +496,11 @@ do
                                         sed -i $hist'd' ~/.encode_file
                                     fi
                                     echo "$to#$file_encode_txt#$hd" >> ~/.encode_file
-                                    echo "rm $init"
-                                    rm "$init"
+                                    if [ "$n" == "0" ] 
+                                    then
+                                        echo "rm $init"
+                                        rm "$init"
+                                    fi
                                 fi
                                 notify-send "convertion de $init terminée en $runtime secondes"
                             else
